requirements.txt
streamlit
pymongo
certifi
plotly
pandas
cryptography
python-dotenv
MetaTrader5